Route to submit page on re-import
