+++ plenario/utils/helpers.py
@@ -57,85 +57,6 @@ def iter_column(idx, f):
     col_type, null_values = normalize_column_type(col)
     return col_type, null_values
 
-def get_socrata_data_info(host, path, four_by_four, is_shapefile=False):
-    errors = []
-    status_code = None
-    dataset_info = {}
-    print host, path, four_by_four
-    view_url = '%s/%s/%s' % (host, path, four_by_four)
-    
-    if is_shapefile:
-        source_url = '%s/download/%s/application/zip' % (host, four_by_four)
-    else:
-        source_url = '%s/rows.csv?accessType=DOWNLOAD' % view_url
-
-    try:
-        r = requests.get(view_url)
-        status_code = r.status_code
-    except requests.exceptions.InvalidURL:
-        errors.append('Invalid URL')
-    except requests.exceptions.ConnectionError:
-        errors.append('URL can not be reached')
-    try:
-        resp = r.json()
-    except AttributeError:
-        errors.append('No Socrata views endpoint available for this dataset')
-        resp = None
-    except ValueError:
-        errors.append('The Socrata dataset you supplied is not available currently')
-        resp = None
-    if resp:
-        dataset_info = {
-            'name': resp['name'],
-            'description': resp.get('description'),
-            'attribution': resp.get('attribution'),
-            'columns': [],
-            'view_url': view_url,
-            'source_url': source_url
-        }
-        try:
-            dataset_info['update_freq'] = \
-                resp['metadata']['custom_fields']['Metadata']['Update Frequency']
-        except KeyError:
-            dataset_info['update_freq'] = None
-
-        columns = resp.get('columns') #presence of columns implies table file
-        if columns:
-            for column in columns:
-                d = {
-                    'human_name': column['name'],
-                    'machine_name': column['fieldName'],
-                    #'field_name': column['fieldName'], # duplicate definition for code compatibility
-                    #'field_name': column['name'], # duplicate definition for code compatibility
-                    'field_name': slugify(column['name']), # duplicate definition for code compatibility
-                    'data_type': column['dataTypeName'],
-                    'description': column.get('description', ''),
-                    'width': column['width'],
-                    'sample_values': [],
-                    'smallest': '',
-                    'largest': '',
-                }
-
-                if column.get('cachedContents'):
-                    cached = column['cachedContents']
-                    if cached.get('top'):
-                        d['sample_values'] = \
-                            [c['item'] for c in cached['top']][:5]
-                    if cached.get('smallest'):
-                        d['smallest'] = cached['smallest']
-                    if cached.get('largest'):
-                        d['largest'] = cached['largest']
-                    if cached.get('null'):
-                        if cached['null'] > 0:
-                            d['null_values'] = True
-                        else:
-                            d['null_values'] = False
-                dataset_info['columns'].append(d)
-        else:
-            if not is_shapefile:
-                errors.append('Views endpoint not structured as expected')
-    return dataset_info, errors, status_code
-
 
 def slugify(text, delim=u'_'):
     """
